agentic-ai-tutor/requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# LLM
google-generativeai==0.8.4

# Embeddings (local)
sentence-transformers==3.3.1
torch>=2.0.0

# Vector DB
chromadb==0.6.3

# Database
aiosqlite==0.20.0

# Text processing
markdown==3.7
beautifulsoup4==4.12.3
tiktoken==0.8.0

# Utilities
structlog==24.4.0
httpx==0.28.1
numpy>=1.26.0
rich==13.9.4
